Crate pipeline

Modules

error

etl 🔒

etl_error

manager 🔒

metrics 🔒

pipeline_operator

table

util

Structs

Array

GreptimeTransformer

fields not in the columns will be discarded to prevent automatic column creation in GreptimeDB

Map

Pipeline

SchemaInfo

This is used to record the current state schema information and a sequential cache of field names. As you traverse the user input JSON, this will change. It will record a superset of all user input schemas.

SelectInfo

SelectInfo is used to store the selected keys from OpenTelemetry record attrs

Enums

Content

PipelineWay

Value

Value can be used as type acts as value: the enclosed value is the actual value acts as type: the enclosed value is the default value

Traits

Processor

Processor trait defines the interface for all processors.

Transformer

Functions

identity_pipeline

Identity pipeline for Greptime This pipeline will convert the input JSON array to Greptime Rows

parse

Type Aliases

PipelineInfo

Pipeline info. A tuple of timestamp and pipeline reference.

PipelineRef

PipelineTableRef

PipelineVersion

Pipeline version. An optional timestamp with nanosecond precision.

Result