Module pipeline::etl::field

Structs

Field

Raw processor-defined inputs and outputs

Fields

A collection of fields.

InputFieldInfo

Information about the input field including the name and index in intermediate keys.

OneInputMultiOutputField

Information about a field that has one input and multiple outputs.

OneInputOneOutputField

Information about a field that has one input and one output.