pipeline::etl::processor
Trait ProcessorBuilder

pub trait ProcessorBuilder:
    Debug
    + Send
    + Sync
    + 'static {
    // Required methods
    fn output_keys(&self) -> HashSet<&str>;
    fn input_keys(&self) -> HashSet<&str>;
    fn build(self, intermediate_keys: &[String]) -> Result<ProcessorKind>;
}

ProcessorBuilder trait defines the interface for all processor builders A processor builder is used to create a processor

Required Methods

fn output_keys(&self) -> HashSet<&str>

Get the processor’s output keys

fn input_keys(&self) -> HashSet<&str>

Get the processor’s input keys

fn build(self, intermediate_keys: &[String]) -> Result<ProcessorKind>

Build the processor