pipeline::etl::transform::transformer

Module greptime

Modules

coerce

Structs

GreptimePipelineParams

Parameters that can be used to configure the greptime pipelines.

GreptimeTransformer

fields not in the columns will be discarded to prevent automatic column creation in GreptimeDB

SchemaInfo

This is used to record the current state schema information and a sequential cache of field names. As you traverse the user input JSON, this will change. It will record a superset of all user input schemas.

Constants

DEFAULT_GREPTIME_TIMESTAMP_COLUMN 🔒

DEFAULT_MAX_NESTED_LEVELS_FOR_JSON_FLATTENING 🔒

Functions

calc_ts 🔒

decide_semantic 🔒

do_flatten_object 🔒

flatten_object

Consumes the JSON object and consumes it into a single-level object.

identity_pipeline

Identity pipeline for Greptime This pipeline will convert the input JSON array to Greptime Rows params table is used to set the semantic type of the row key column to Tag

identity_pipeline_inner 🔒

resolve_number_schema 🔒

resolve_schema 🔒

resolve_value 🔒

values_to_row 🔒