pipeline::etl::transform::transformer::greptime
Function flatten_json_object

pub fn flatten_json_object(
    object: Map<String, Value>,
    max_nested_levels: usize,
) -> Result<Map<String, Value>>

Consumes the JSON object and consumes it into a single-level object.

The max_nested_levels parameter is used to limit the nested levels of the JSON object. The error will be returned if the nested levels is greater than the max_nested_levels.