pipeline::etl::value

Enum Value

pub enum Value {

    Null,
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Uint8(u8),
    Uint16(u16),
    Uint32(u32),
    Uint64(u64),
    Float32(f32),
    Float64(f64),
    Boolean(bool),
    String(String),
    Timestamp(Timestamp),
    Array(Array),
    Map(Map),

}

Value can be used as type acts as value: the enclosed value is the actual value acts as type: the enclosed value is the default value

Variants

Null

Int8(i8)

Int16(i16)

Int32(i32)

Int64(i64)

Uint8(u8)

Uint16(u16)

Uint32(u32)

Uint64(u64)

Float32(f32)

Float64(f64)

Boolean(bool)

String(String)

Timestamp(Timestamp)

Array(Array)

We only consider object and array to be json types.

Map(Map)

Implementations

impl Value

pub fn is_null(&self) -> bool

pub fn parse_str_type(t: &str) -> Result<Self>

pub fn parse_str_value(&self, v: &str) -> Result<Self>

only support string, bool, number, null

pub fn to_str_value(&self) -> String

only support string, bool, number, null

pub fn to_str_type(&self) -> &str

pub fn get(&self, key: &str) -> Option<&Self>

pub fn as_str(&self) -> Option<&str>

pub fn as_i64(&self) -> Option<i64>

pub fn as_u64(&self) -> Option<u64>

pub fn as_f64(&self) -> Option<f64>

pub fn pointer(&self, pointer: &str) -> Option<&Value>

pub fn pointer_mut(&mut self, pointer: &str) -> Option<&mut Value>