pipeline

Module manager

Modules

pipeline_cache 🔒

pipeline_operator

table

util

Structs

PipelineContext

SelectInfo

SelectInfo is used to store the selected keys from OpenTelemetry record attrs The key is used to uplift value from the attributes and serve as column name in the table

Enums

IdentityTimeIndex

PipelineDefinition

Enum for holding information of a pipeline, which is either pipeline itself, or information that be used to retrieve a pipeline from PipelineHandler

PipelineWay

Constants

GREPTIME_INTERNAL_IDENTITY_PIPELINE_NAME

GREPTIME_INTERNAL_TRACE_PIPELINE_V0_NAME

GREPTIME_INTERNAL_TRACE_PIPELINE_V1_NAME

IDENTITY_TS_DATESTR 🔒

IDENTITY_TS_EPOCH 🔒

Functions

if_ignore_errors 🔒

time_unit_to_value_data 🔒

Type Aliases

PipelineInfo

Pipeline info. A tuple of timestamp and pipeline reference.

PipelineRef

PipelineTableRef

PipelineVersion

Pipeline version. An optional timestamp with nanosecond precision.