Crate pipeline

Modules

error

etl 🔒

manager 🔒

pipeline_operator

table

Structs

GreptimeTransformer

fields not in the columns will be discarded to prevent automatic column creation in GreptimeDB

Pipeline

Enums

Content

Value

Value can be used as type acts as value: the enclosed value is the actual value acts as type: the enclosed value is the default value

Functions

parse