pipeline::etl
Module value

Re-exports

pub use array::Array;

pub use map::Map;

pub use time::Timestamp;

Modules

array

map

time

Enums

Value

Value can be used as type acts as value: the enclosed value is the actual value acts as type: the enclosed value is the default value