Trait pipeline::Processor

pub trait Processor: Debug + Send + Sync + 'static {
    // Required methods
    fn fields(&self) -> &Fields;
    fn fields_mut(&mut self) -> &mut Fields;
    fn kind(&self) -> &str;
    fn ignore_missing(&self) -> bool;
    fn output_keys(&self) -> HashSet<String>;
    fn exec_field(&self, val: &Value, field: &Field) -> Result<Map, String>;
    fn exec_mut(&self, val: &mut Vec<Value>) -> Result<(), String>;

    // Provided method
    fn exec_map(&self, map: &mut Map) -> Result<(), String> { ... }
}

Processor trait defines the interface for all processors A processor is a transformation that can be applied to a field in a document It can be used to extract, transform, or enrich data Now Processor only have one input field. In the future, we may support multiple input fields. The output of a processor is a map of key-value pairs that will be merged into the document when you use exec_map method.

Required Methods

fn fields(&self) -> &Fields

Get the processor’s fields fields is just the same processor for multiple keys. It is not the case that a processor has multiple inputs

fn fields_mut(&mut self) -> &mut Fields

Get the processor’s fields mutably

fn kind(&self) -> &str

Get the processor’s kind

fn ignore_missing(&self) -> bool

Whether to ignore missing

fn output_keys(&self) -> HashSet<String>

processor all output keys if a processor has multiple output keys, it should return all of them

fn exec_field(&self, val: &Value, field: &Field) -> Result<Map, String>

Execute the processor on a document and return a map of key-value pairs

fn exec_mut(&self, val: &mut Vec<Value>) -> Result<(), String>

Execute the processor on a vector which be preprocessed by the pipeline

Provided Methods

fn exec_map(&self, map: &mut Map) -> Result<(), String>

Execute the processor on a map and merge the output into the original map